suspension_setup_evaluator/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web Interface
streamlit>=1.28.0

# Configuration
PyYAML>=6.0

# Utils
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
